+++ shello_cli/__init__.py
@@ -27,5 +27,5 @@ try:
 except Exception:
     pass
 
-__version__ = "0.8.3"
+__version__ = "0.8.4"